refactor: split symbols.rs into focused modules
Extract per-language symbol extractors and function signature rewriting
from the monolithic symbols.rs (3001 lines) into dedicated modules:

- ast/symbol_extract.rs (595 lines): per-language extractors for 20
  languages (visit_node, extract_rust, extract_python, extract_go, etc.)
- ast/rewrite.rs (1086 lines): FunctionSpan, FunctionSigEdit,
  find_function_span, rewrite_function_signature, and all helpers
- ast/symbols.rs (1355 lines, was 3001): core types (SymbolKind,
  SymbolDef), entry points, span/filter utilities, re-exports

Re-exports in symbols.rs ensure all external imports continue to work
without changes to any file outside src/ast/.

Closes #1358
